# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "api/Managers/getById/{id}": {
            "get": {
                "tags": [
                    "Managers"
                ],
                "summary": "Get single Managers",
                "description": "Returns Get single Managers",
                "operationId": "GetSingleManager",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Managers id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Manager"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/auth/Managers/create": {
            "post": {
                "tags": [
                    "Managers"
                ],
                "summary": "Insert new Manager",
                "description": "Returns Manager data",
                "operationId": "Insert Manager",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Manager"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Manager"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "api/auth/Managers/{id}/update": {
            "post": {
                "tags": [
                    "Managers"
                ],
                "summary": "Insert new Manager info",
                "description": "Returns Manager data",
                "operationId": "Update Manager",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Manager id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Manager"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Manager"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "api/auth/Managers/{id}/delete": {
            "get": {
                "tags": [
                    "Managers"
                ],
                "summary": "Delete single Manager",
                "description": "Returns Deleted",
                "operationId": "DeleteSingleManager",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Manager id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Manager"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
        "api/auth/Users/{id}/update": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Insert new User info",
                "description": "Returns User data",
                "operationId": "Update User",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
        "api/Users/{id}/getById": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get single Users",
                "description": "Returns Get single Users",
                "operationId": "GetSingleUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Users id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/auth/Users/{id}/delete": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete single User",
                "description": "Returns Deleted",
                "operationId": "DeleteSingleUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Manager": {
                "title": "Manager",
                "description": "Manager model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Manager name",
                        "type": "string",
                        "example": "Sara"
                    },
                    "phone_number": {
                        "title": "phone",
                        "description": "Manager phone",
                        "type": "string",
                        "example": "0770*******"
                    },
                    "img": {
                        "title": "img",
                        "description": "Manager picture",
                        "type": "string",
                        "example": "img"
                    },
                    "email": {
                        "title": "email",
                        "description": "email of the new user",
                        "type": "string",
                        "example": "sara@gmail.com"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Manager"
                }
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "name of the new user",
                        "type": "string",
                        "example": "sara"
                    },
                    "type": {
                        "title": "type",
                        "description": "type of the new user",
                        "type": "string",
                        "example": "admin"
                    },
                    "phone_number": {
                        "title": "phone_number",
                        "description": "phone_number of the new user",
                        "type": "string",
                        "example": "07707722504"
                    },
                    "password": {
                        "title": "password",
                        "description": "Password of the new user",
                        "type": "string",
                        "example": "sara1234"
                    },
                    "email": {
                        "title": "email",
                        "description": "email of the new user",
                        "type": "string",
                        "example": "sara@gmail.com"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}